refactor(frontend): move Loader to elements folder
