+++ packages/frontend/src/ProtectedRoute.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react';
 import { Redirect, Route } from 'react-router';
 import Cookies from 'universal-cookie';
 import { useAuth } from './hooks';
-import { Loader } from './components/atoms';
+import { Loader } from './components/elements';
 
 const ProtectedRoute = ({
   path,
