+++ packages/frontend/src/components/atoms/index.ts
@@ -1,7 +1,6 @@
-import Loader from './Loader';
 import LogOut from './LogOut';
 import UserInput from './UserInput';
 import UserSubmit from './UserSubmit';
 import MiniProduct from './MiniProduct';
 
-export { Loader, LogOut, UserInput, UserSubmit, MiniProduct };
+export { LogOut, UserInput, UserSubmit, MiniProduct };
