+++ packages/frontend/src/components/containers/LoginForm.tsx
@@ -7,7 +7,7 @@ import Cookie from 'universal-cookie';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-import { Loader } from '../atoms';
+import { Loader } from '../elements';
 import { Input } from '../elements';
 import { HTTPException } from '../../utils/HttpException';
