+++ packages/frontend/src/components/containers/PayPal.tsx
@@ -6,7 +6,7 @@ import Cookie from 'universal-cookie';
 import { toast } from 'react-toastify';
 
 import { API_KEY, API_URI, MODE } from '../../config';
-import { Loader } from '../atoms';
+import { Loader } from '../elements';
 import { IProduct } from '../../context/interfaces';
 
 const PayPal = ({ products }: IPayPalProps) => {
