+++ packages/frontend/src/components/elements/index.ts
@@ -1,4 +1,5 @@
 import Nav from './Nav';
 import Input from './Input';
+import Loader from './Loader';
 
-export { Nav, Input };
+export { Nav, Input, Loader };
